Resolve type "overflow" false negative in safe casting check
 - Move type_traits to separate header
